requirements.txt
apiai==1.2.3
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
dialogflow==0.5.1
Django==2.0.2
django-shell-plus==1.1.7
factory-boy==2.11.1
Faker==0.9.1
google-api-core==1.4.1
google-auth==1.5.1
google-cloud==0.34.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
idna==2.7
numpy==1.15.2
protobuf==3.6.1
psycopg2==2.7.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.3
python-slugify==1.2.6
pytz==2018.5
requests==2.19.1
rsa==4.0
six==1.11.0
text-unidecode==1.2
Unidecode==1.0.22
urllib3==1.23
django-extensions
celery